feat: rename learning method to small box
