+++ lib/features/landing_page/presentation/widgets/small_box.dart
@@ -1,0 +1,80 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:responsive_sizer/responsive_sizer.dart';
+
+class SmallBox extends ConsumerWidget {
+  final String title;
+  final String description;
+  final VoidCallback? onLearnMore;
+
+  const SmallBox(
+      {required this.title,
+      required this.description,
+      this.onLearnMore,
+      super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Flexible(
+      child: Card(
+        clipBehavior: Clip.antiAliasWithSaveLayer,
+        color: Theme.of(context).cardColor,
+        elevation: 0,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(8),
+        ),
+        child: Padding(
+          padding: const EdgeInsetsDirectional.fromSTEB(10, 10, 10, 10),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisSize: MainAxisSize.max,
+            children: [
+              Text(title, style: Theme.of(context).textTheme.titleSmall),
+              Text(
+                description,
+                style: TextStyle(fontSize: 14.sp),
+                maxLines: 2,
+                overflow: TextOverflow.ellipsis,
+              ),
+              const SizedBox(height: 5),
+              onLearnMore != null
+                  ? Row(
+                      mainAxisSize: MainAxisSize.max,
+                      children: [
+                        TextButton(
+                            onPressed: onLearnMore,
+                            style: TextButton.styleFrom(
+                              minimumSize: Size.zero,
+                              padding: const EdgeInsets.fromLTRB(1, 3, 1, 3),
+                              tapTargetSize: MaterialTapTargetSize.shrinkWrap,
+                            ),
+                            child: Row(
+                              children: [
+                                Text('Learn More  ',
+                                    style: Theme.of(context)
+                                        .textTheme
+                                        .bodyMedium
+                                        ?.copyWith(
+                                            color: Theme.of(context)
+                                                .colorScheme
+                                                .secondary)),
+                                Icon(
+                                  Icons.arrow_forward,
+                                  color: Theme.of(context)
+                                      .textTheme
+                                      .headlineLarge
+                                      ?.color,
+                                  size: 14,
+                                ),
+                              ],
+                            )),
+                      ],
+                    )
+                  : SizedBox(width: 0),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
